refactore: Cria instâncias onde o type Purchase era utilizado.
