+++ modulo6/classes-encapsulamento/src/class/Purchase.ts
@@ -1,3 +1,7 @@
+// --------------------------------Exercício 3 ------------------------------------ //
+// Por fim, crie a classe Purchase no lugar de seu type. Não se preocupe com o refatoramento e uso do PurchaseDB. 
+// Finalize a implementação com o uso de instâncias onde o type Purchase era utilizado.
+
 export class Purchase {
     private id: string
     private userId: string
@@ -5,6 +9,7 @@ export class Purchase {
     private quantity: number
     private totalPrice: number
 
+    // Construtor
     constructor(id: string, userId: string, ProductId: string, quantity: number, totalPrice: number) {
         this.id = id;
         this.userId = userId;
@@ -13,6 +18,7 @@ export class Purchase {
         this.totalPrice = totalPrice;
     };
 
+    // Método de acesso
     public getId(): string {
         return this.id;
     };
@@ -29,6 +35,7 @@ export class Purchase {
         return this.totalPrice;
     };
 
+    // Método de alteração ou definição de atributos
     public setId(newId: string): void {
         this.id = newId;
     };
