+++ modulo6/classes-encapsulamento/src/endpoints/createPurchase.ts
@@ -1,11 +1,13 @@
-import { Request, Response } from "express"
-import connection from "../database/connection"
-import { TABLE_PRODUCTS, TABLE_PURCHASES, TABLE_USERS } from "../database/tableNames"
-import { TProduct } from "../models/Product"
-import { Purchase } from "../models/Purchase"
+import { Request, Response } from "express";
+import { TABLE_PRODUCTS, TABLE_PURCHASES, TABLE_USERS } from "../database/tableNames";
+import { Product } from "../class/Product";
+import { Purchase } from "../class/Purchase";
+import connection from "../database/connection";
 
 export const createPurchase = async (req: Request, res: Response) => {
+
     let errorCode = 400
+
     try {
         const userId = req.body.userId
         const productId = req.body.productId
@@ -33,29 +35,43 @@ export const createPurchase = async (req: Request, res: Response) => {
             throw new Error("Produto não encontrado.")
         }
 
-        const product: TProduct = {
-            id: findProduct[0].id,
-            name: findProduct[0].name,
-            price: findProduct[0].price
-        }
+        // const product: Product = {
+        //     id: findProduct[0].id,
+        //     name: findProduct[0].name,
+        //     price: findProduct[0].price
+        // }
 
-        const newPurchase: Purchase = {
-            id: Date.now().toString(),
+        const product = new Product(
+            findProduct[0].id,
+            findProduct[0].name,
+            findProduct[0].price,
+        )
+
+        // const newPurchase: Purchase = {
+        //     id: Date.now().toString(),
+        //     userId,
+        //     productId,
+        //     quantity,
+        //     totalPrice: product.price * quantity
+        // }
+
+        const purchase = new Purchase(
+            Date.now().toString(),
             userId,
             productId,
             quantity,
-            totalPrice: product.price * quantity
-        }
+            product.getPrice() * quantity
+        )
 
         await connection(TABLE_PURCHASES).insert({
-            id: newPurchase.id,
-            user_id: newPurchase.userId,
-            product_id: newPurchase.productId,
-            quantity: newPurchase.quantity,
-            total_price: newPurchase.totalPrice
+            id: purchase.getId(),
+            user_id: purchase.getUserId(),
+            product_id: purchase.getProductId(),
+            quantity: purchase.getQuantity(),
+            total_price: purchase.getTotalPrice()
         })
 
-        res.status(201).send({ message: "Compra registrada", purchase: newPurchase })
+        res.status(201).send({ message: "Compra registrada", purchase: purchase })
     } catch (error) {
         res.status(errorCode).send({ message: error.message })
     }
